#ca-13 fixed

+++ lib/centos/rating/patches/issues_controller.rb
@@ -13,31 +13,14 @@ module Centos::Rating::Patches
     module InstanceMethods
       protected
       def update_rating
-        pars = params[:issue][:rating]
-        user = ::User.current
-        rating = ::Rating.issue_rating @issue, user
-
-        return unless !pars.blank? && user.allowed_to?(:centos_rate, @issue.project)
-
-        if rating
-          if pars[:mark].blank?
-            rating.destroy
-          else
-            rating.safe_attributes = pars
-            rating.save
-          end
-        else
-          unless pars[:mark].blank?
-            ::Rating.create({
-              issue_id: @issue.id,
-              project_id: @issue.project_id,
-              evaluator_id: user.id,
-              evaluated_id: @issue.assigned_to_id,
-              mark: pars[:mark],
-              comments: pars[:comments]
-            }) 
-          end
-        end
+        rating_from_params = params[:issue][:rating]
+        return unless !rating_from_params.blank? && ::User.current.allowed_to?(:centos_rate, @issue.project)
+        rating = ::Rating.find_or_initialize_by issue_id: @issue.id,
+                                                evaluator_id: ::User.current.id,
+                                                evaluated_id: rating_from_params.delete(:evaluated_id)
+        return rating.destroy if rating_from_params[:mark].blank?
+        rating.safe_attributes = rating_from_params
+        rating.save
       end
     end
   end
